Refactor: view, store,interface 분리해서 main, root 리팩토링
